+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 5.0 [not yet released]
 
-- the artificial tag duration:seconds is now a long, not a string
+- request gzipping for matrix and route clients (#2511)
+- bugfix: client-hc now considers headings and custom models (#2009, #2535)
+- the artificial tag duration:seconds is now a long, no longer a string, commit 6d81d8ae8de52987522991edd835e42c8d2046cf
 - added FlagEncoder#getName (use just like toString() before), commit 86f6a8b5209ad8ef47c24d935f5746e7694eb11c
 - faster edge-based CH preparation, especially with large u-turn costs and GermanyCountryRule (many large weight edges
   due to access=destination on tracks) (#2522)